requirements.txt
alembic==1.1.0
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
elasticsearch==7.0.4
entrypoints==0.3
eventsourcing==7.1.6
flake8==3.7.8
Flask==1.1.1
Flask-BabelEx==0.9.3
flask-buzz==0.1.14
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-mongoengine==0.9.5
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
langcodes==1.4.1
Mako==1.1.0
marisa-trie==0.7.5
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.18.2
mongomock==3.18.0
more-itertools==7.2.0
mysql-connector==2.2.9
packaging==19.2
passlib==1.7.1
pendulum==2.0.5
pluggy==0.13.0
py==1.8.0
py-buzz==0.3.7
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.8.2
pyflakes==2.1.1
PyJWT==1.7.1
pymongo==3.9.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.2.1
pytest-ordering==0.6
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
pytzdata==2019.3
readerwriterlock==1.0.4
requests==2.22.0
sentinels==1.0.0
sentry-sdk==0.13.1
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.8
SQLAlchemy-Utils==0.34.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.6
WTForms==2.2.1
zipp==0.6.0
